pyproject.toml
[tool.poetry]
name = "requestforward"
version = "0.1.0"
description = ""
package-mode = false
authors = ["HeMOu <fangqichenchao@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
pyqt6 = "^6.7.1"
configparser = "^7.1.0"
httpx = "^0.28.1"
uvicorn = "^0.32.1"
pyyaml = "^6.0.2"
qasync = "^0.27.1"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
